cleanups in the lintpack package
- Add missing doc comments
- Move all public (exported) API to a single file (lintpack.go)
- Remove unused and confusing+undocummented Checker.Init field
