+++ doc.go
@@ -1,0 +1,5 @@
+// Package lintpack provides shared API between the linter and its checkers.
+//
+// Linter is usually implemented by creating instances of registered checkers.
+// Checkers are registered by the AddChecker call.
+package lintpack
